Swap form views and container file name
