+++ src/components/Forms/NewEventForm/NewEventForm.js
@@ -1,77 +1,140 @@
 import React from 'react'
-import * as Yup from "yup"
-import { Formik } from "formik";
-import NewEventFormContainer from './NewEventFormContainer'
-import { fetchBackend } from '../../../utils'
+import Button from "@material-ui/core/Button";
+import TextField from "@material-ui/core/TextField";
+import Paper from "@material-ui/core/Paper";
+import Typography from "@material-ui/core/Typography"
+import DateFnsUtils from '@date-io/date-fns';
+import {
+    MuiPickersUtilsProvider,
+    KeyboardDateTimePicker
+} from '@material-ui/pickers';
+import ThemeProvider from '../../ThemeProvider'
+const slugify = require('slugify')
 
-export default function NewEventForm() {
+export default function NewEventForm(props) {
+    const {
+        values: { ename, slug, description, capacity, elocation, imageUrl, startDate, endDate },
+        errors,
+        touched,
+        handleSubmit,
+        handleChange,
+        setFieldValue,
+        setFieldTouched
+    } = props;
 
-    const validationSchema = Yup.object({
-        ename: Yup.string().required(),
-        slug: Yup.string().matches(/^[a-z\-0-9]*$/, "Slug must be lowercase and have no whitespace").required(),
-        description: Yup.string().required(),
-        capacity: Yup.number('Valid number required')
-            .min(0, 'Valid capacity required')
-            .required(),
-        // partners: Yup.string().required(),
-        elocation: Yup.string().required(),
-        imageUrl: Yup.string().url().required(),
-    });
-
-    const initialValues = {
-        ename: "",
-        slug: "",
-        description: "",
-        capacity: "",
-        partners: "",
-        elocation: "",
-        imageUrl: "",
-        startDate: new Date(),
-        endDate: new Date()
+    const change = (name, e) => {
+        e.persist();
+        handleChange(e);
+        setFieldTouched(name, true, false);
     };
 
-    return (
-        <Formik
-            initialValues={initialValues}
-            validationSchema={validationSchema}
-            onSubmit={submitValues}
-        >
-            {props => <NewEventFormContainer {...props} />}
-        </Formik>
-    )
-
-    async function submitValues(values) {
-        const body = JSON.stringify({
-            ename: values.ename,
-            id: values.slug,
-            description: values.description,
-            capac: values.capacity,
-            elocation: values.elocation,
-            imageUrl: values.imageUrl,
-            startDate: values.startDate,
-            endDate: values.endDate
-        })
-        fetchBackend('/events/get', 'GET')
-            .then((response) => response.json())
-            .then((response) => {
-                const isDuplicate = response.find(event => event.id === values.slug)
-                if (isDuplicate) {
-                    alert('Event with that slug already exists!')
-                } else {
-                    fetchBackend('/events/create', 'POST', body)
-                        .then((response) => response.json())
-                        .then((response) => {
-                            console.log(response)
-                            alert(response.message)
-                            window.location.href = "/";
-                        })
-                        .catch(err => {
-                            console.log(err)
-                            alert(err.message + ' Please contact a dev')
-                        })
-                }
-            })
-
+    const handleEventNameChange = (name, e) => {
+        e.persist();
+        const newSlug = slugify(e.target.value, { lower: true });
+        setFieldValue('slug', newSlug)
+        handleChange(e);
+        setFieldTouched(name, true, false);
     }
 
+    const handleStartDateChange = (date) => {
+        setFieldValue("startDate", date)
+    }
+
+    const handleEndDateChange = (date) => {
+        setFieldValue("endDate", date)
+    }
+
+    return (
+        <ThemeProvider>
+            <Typography variant="h1">Create a New Event</Typography>
+            <form onSubmit={handleSubmit}>
+                <Paper>
+                    <TextField
+                        id="ename"
+                        label="Event Name"
+                        fullWidth
+                        helperText={touched.ename ? errors.ename : ""}
+                        error={touched.ename && Boolean(errors.ename)}
+                        value={ename}
+                        onChange={handleEventNameChange.bind(null, "ename")}
+                    />
+                    <TextField
+                        id="slug"
+                        label="Slug"
+                        fullWidth
+                        helperText={touched.slug ? errors.slug : ""}
+                        error={touched.slug && Boolean(errors.slug)}
+                        value={slug}
+                        onChange={change.bind(null, "slug")}
+                    />
+                    <TextField
+                        id="description"
+                        label="Description"
+                        multiline
+                        fullWidth
+                        helperText={touched.description ? errors.description : ""}
+                        error={touched.description && Boolean(errors.description)}
+                        value={description}
+                        onChange={change.bind(null, "description")}
+                    />
+                    <TextField
+                        id="capacity"
+                        label="Capacity"
+                        type="number"
+                        min="0"
+                        helperText={touched.capacity ? errors.capacity : ""}
+                        error={touched.capacity && Boolean(errors.capacity)}
+                        value={capacity}
+                        onChange={change.bind(null, "capacity")}
+                    />
+                    {/* <TextField
+                        id="partners"
+                        label="Partners & Sponsors"
+                        fullWidth
+                        helperText={touched.partners ? errors.partners : ""}
+                        error={touched.partners && Boolean(errors.partners)}
+                        value={partners}
+                        onChange={change.bind(null, "partners")}
+                    /> */}
+                    <MuiPickersUtilsProvider utils={DateFnsUtils}>
+                        <KeyboardDateTimePicker
+                            margin="normal"
+                            label="Start Date"
+                            minDate={new Date()}
+                            value={startDate}
+                            onChange={handleStartDateChange}
+                        />
+                        <KeyboardDateTimePicker
+                            margin="normal"
+                            label="End Date"
+                            minDate={startDate}
+                            value={endDate}
+                            onChange={handleEndDateChange}
+                        />
+                    </MuiPickersUtilsProvider>
+                    <TextField
+                        id="elocation"
+                        label="Location"
+                        fullWidth
+                        helperText={touched.elocation ? errors.elocation : ""}
+                        error={touched.elocation && Boolean(errors.elocation)}
+                        value={elocation}
+                        onChange={change.bind(null, "location")}
+                    />
+                    <TextField
+                        id="imageUrl"
+                        label="Image URL"
+                        fullWidth
+                        helperText={touched.imageUrl ? errors.imageUrl : ""}
+                        error={touched.imageUrl && Boolean(errors.imageUrl)}
+                        value={imageUrl}
+                        onChange={change.bind(null, "imageUrl")}
+                    />
+                </Paper>
+                <Button type="submit">
+                    Submit
+                </Button>
+            </form>
+        </ThemeProvider>
+    )
 }
