--- src/components/Forms/index.js
@@ -1,3 +1,0 @@
-import { NewEventForm } from './NewEventForm'
-
-export { NewEventForm }
