+++ src/components/ScrollToTop.js
@@ -1,6 +1,8 @@
 import { useEffect } from "react";
 import { useLocation } from "react-router-dom";
 
+// Helper component to scroll to top
+// React router doesn't automatically scroll to top when navigating to a different route
 export default function ScrollToTop() {
     const { pathname } = useLocation();
